+++ dbscripts/dblib.py
@@ -1,5 +1,6 @@
+import copy
 import re
-
+import logging
 import psycopg as pg
 from envex import Env
 
@@ -14,106 +15,146 @@ __all__ = (
 )
 
 
-class UnconfiguredEnvironment(BaseException):
+class EnvironmentNotConfigured(BaseException):
     pass
 
 
 env = Env(readenv=True, parents=True)
 
 
-def pg_db_info(host=None, port=None, name=None, role=None, user=None, pswd=None, url=None) -> dict:
-    db = {
-        "scheme": "postgres",
-        "host": host or env("DBHOST"),
-        "port": port or env("DBPORT"),
-        "name": name or env("DBNAME"),
-        "user": role or env("DBUSER"),
-        "role": user or env("DBROLE", default=env("DBUSER")),
-        "pass": pswd or env("DBPASS"),
-    }
-    if db["host"] is not None and ':' in db["host"]:
-        db["host"], db["port"] = db["host"].split(':', maxsplit=1)
-    if not url and env.is_set("DATABASE_URL"):
-        url = env("DATABASE_URL")
-    if url:  # parse it
-        p = re.compile(r"""
-                       (?P<scheme>[\w+]+)://
-                       (?:(?P<user>[^:/]*)(?::(?P<pswd>[^@]*))?@)?
-                       (?:(?:\[(?P<ipv6host>[^/?]+)]|(?P<ipv4host>[^/:?]+))?(?::(?P<port>[^/?]*))?)?
-                       (?:/(?P<name>[^?]*))?(?:\?(?P<query>.*))?""",
-                       re.X,)
+class DBUrl:
 
-        if (m := p.match(url)) is not None:
-            for k, v in m.groupdict().items():
-                db[k] = v
-            db["host"] = db["ipv4host"] or db["ipv6host"]
+    def __init__(self, *args, host=None, port=None, name=None, role=None, user=None, pswd=None, url=None, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.scheme = "postgresql"
+        self.host = env("DBHOST")
+        self.port = env("DBPORT")
+        self.name = env("DBNAME")
+        self.user = env("DBUSER")
+        self.role = env("DBROLE", default=env("DBUSER"))
+        self.pswd = env("DBPASS")
+        self.ipv4host = self.ipv6host = ''
 
-    db["url"] = f"{db['scheme']}://{db['user']}:{db['pass']}@{db['host']}:{db['port'] or 5432}/{db['name']}"
-    return db
+        if self.host and ':' in self.host:
+            self.host, self.port = self.host.split(':', maxsplit=1)
+        if not url and env.is_set("DATABASE_URL"):
+            url = env("DATABASE_URL")
+        if url:  # parse it
+            p = re.compile(r"""
+                           (?P<scheme>[\w+]+)://
+                           (?:(?P<user>[^:/]*)(?::(?P<pswd>[^@]*))?@)?
+                           (?:(?:\[(?P<ipv6host>[^/?]+)]|(?P<ipv4host>[^/:?]+))?(?::(?P<port>[^/?]*))?)?
+                           (?:/(?P<name>[^?]*))?(?:\?(?P<query>.*))?""",
+                           re.X, )
+
+            if (m := p.match(url)) is not None:
+                for k, v in m.groupdict().items():
+                    setattr(self, k, v)
+                self.host = self.ipv4host or self.ipv6host
+
+        # allow overrides
+
+        if host:
+            if ':' in host:
+                self.host, self.port = self.host.split(':', maxsplit=1)
+            else:
+                self.host = host
+        if port:
+            self.port = port
+        if name:
+            self.name = name
+        if user:
+            self.user = user
+            if not role:
+                self.role = user
+        if pswd:
+            self.pswd = pswd
+
+    def to_dict(self):
+        return dict(scheme=self.scheme, host=self.host, port=self.port,
+                    name=self.name, user=self.user, role=self.role, pswd=self.pswd)
+
+    def url(self):
+        return f"{self.scheme}://{self.user}:{self.pswd}@{self.host}:{self.port or 5432}/{self.name}"
+
+    def __getitem__(self, value):
+        if hasattr(self, value):
+            return getattr(self, value)
+        raise KeyError(value)
+
+    def copy(self):
+        return copy.copy(self)
 
 
-def pg_dsn(db, sa=False):
-    dsn = env('SA_DATABASE_URL') if sa else db['url']
+def pg_db_info(host=None, port=None, name=None, role=None, user=None, pswd=None, url=None) -> DBUrl:
+    return DBUrl(host=host, port=port, name=name, role=role, user=user, pswd=pswd, url=url)
+
+
+def pg_dsn(db: DBUrl, sa: bool = False) -> str:
+    dsn = env('SA_DATABASE_URL') if sa else db.url()
     if not dsn:
-        raise UnconfiguredEnvironment('SA_DATABASE_URL' if sa else 'DATABASE_URL')
+        raise EnvironmentNotConfigured('SA_DATABASE_URL' if sa else 'DATABASE_URL')
     return dsn
 
 
-def pg_connect(db, sa=False, **kwargs):
+def pg_connect(db: DBUrl, sa: bool = False, **kwargs):
     conn = pg.connect(pg_dsn(db, sa=sa), **kwargs)
     conn.autocommit = True
     return conn
 
 
-def pg_clear_connections(db, **kwargs):
+def pg_clear_connections(db: DBUrl, **kwargs) -> None:
+    if db.name:
+        conn = pg_connect(db, sa=True, **kwargs)
+        with conn.cursor() as cursor:
+            cursor.execute(f"""SELECT pg_terminate_backend(pid) """
+                           f"""FROM postgres.pg_catalog.pg_stat_activity """
+                           f"""WHERE datname='{db.name}'""")
+        conn.close()
+
+
+def pg_drop_database(db: DBUrl, **kwargs) -> None:
     conn = pg_connect(db, sa=True, **kwargs)
     with conn.cursor() as cursor:
-        cursor.execute(f"""SELECT pg_terminate_backend(pid) """
-                       f"""FROM postgres.pg_catalog.pg_stat_activity """
-                       f"""WHERE datname='{db["name"]}'""")
+        cursor.execute(f"""DROP DATABASE IF EXISTS {db.name}""")
+        cursor.execute(f"""DROP USER IF EXISTS {db.user}""")
+        if db.role and db.role != db.user:
+            cursor.execute(f"""DROP ROLE IF EXISTS {db.role}""")
     conn.close()
+    logging.info(f"""Database '{db.name}' dropped""")
 
 
-def pg_drop_database(db, **kwargs):
-    conn = pg_connect(db, sa=True, **kwargs)
-    with conn.cursor() as cursor:
-        cursor.execute(f"""DROP DATABASE IF EXISTS {db["name"]}""")
-        cursor.execute(f"""DROP USER IF EXISTS {db["user"]}""")
-        cursor.execute(f"""DROP ROLE IF EXISTS {db["role"]}""")
-    conn.close()
-    print(f"""Database '{db["name"]}' dropped""")
-
-
-def pg_database_exists(db, silent=False, **kwargs):
-    test_exc = None
+def pg_database_exists(db: DBUrl, silent: bool = False, raise_error: bool = False, **kwargs):
     try:
         conn = pg_connect(db, sa=False, **kwargs)
         conn.close()
         if not silent:
-            print(f"""Database '{db["name"]}' exists""")
+            logging.info(f"""Database '{db.name}' exists""")
         return True
     except pg.errors.DatabaseError as exc:
-        test_exc = exc
+        if raise_error:
+            raise
+        if not silent:
+            logging.warning(f"""Database '{db.name}': {exc}""")
     if not silent:
-        if test_exc:
-            print(f"""Database '{db["name"]}': {test_exc}""")
-        print(f"""Database '{db["name"]}' does not yet exist""")
+        logging.info(f"""Database '{db.name}' does not yet exist or credentials are incorrect""")
     return False
 
 
-def pg_setup(db, **kwargs):
+def pg_setup(db: DBUrl, **kwargs):
     conn = pg_connect(db, sa=True, **kwargs)
     with conn.cursor() as cursor:
-        cursor.execute(f"""CREATE ROLE {db["role"]}""")
-        cursor.execute(f"""CREATE USER {db["user"]} CREATEDB INHERIT PASSWORD '{db["pass"]}'""")
-        cursor.execute(f"""GRANT {db["role"]} to {db["user"]}""")
-        cursor.execute(f"""ALTER ROLE {db["role"]} SET client_encoding to 'utf8'""")
-        cursor.execute(f"""ALTER ROLE {db["role"]} SET default_transaction_isolation to 'read committed'""")
-        cursor.execute(f"""ALTER ROLE {db["role"]} SET timezone to 'UTC'""")
-    print(f"""Database '{db["name"]}' created""")
+        cursor.execute(f"""CREATE USER {db.user} CREATEDB INHERIT PASSWORD '{db.pswd}'""")
+        if db.role != db.user:
+            cursor.execute(f"""CREATE ROLE {db.role}""")
+            cursor.execute(f"""GRANT {db.role} to {db.user}""")
+        cursor.execute(f"""ALTER ROLE {db.role} SET client_encoding to 'utf8'""")
+        cursor.execute(f"""ALTER ROLE {db.role} SET default_transaction_isolation to 'read committed'""")
+        cursor.execute(f"""ALTER ROLE {db.role} SET timezone to 'UTC'""")
+    logging.info(f"""Database '{db.name}' created""")
 
     with conn.cursor() as cursor:
-        cursor.execute(f"""CREATE DATABASE {db["name"]} WITH OWNER {db["role"]}""")
-        cursor.execute(f"""GRANT ALL PRIVILEGES ON DATABASE {db["name"]} TO {db["role"]}""")
+        cursor.execute(f"""CREATE DATABASE {db.name} WITH OWNER {db.role}""")
+        cursor.execute(f"""GRANT ALL PRIVILEGES ON DATABASE {db.name} TO {db.role}""")
 
     conn.close()