+++ ChangeLog.md
@@ -3,3 +3,8 @@
 - Bug fixes for db exists
 ## 1.2.0
 - change driver to psycopg[binary]
+## 1.3.0
+- change from poetry to hatch
+- add dbready.py
+- changed connection logic to allow --host / --port etc to override URL
+- print -> logging.info